# optimization of max_pool3d forward