<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs
### Describe
<!-- Describe what this PR does -->
* Environment:
    * V100-32G, CUDA 11.2, cuDNN 8
* Feature：
    * replace the div and mod operation with fast_divmod operation
    * replace 1d gpu launch with 2d gpu launch
* Performance (OP Benchmark):

| Paddle Kernel | Config ID | Performance Before | Performance After | Improvement |
| :-----| :---- | :---- | :---- | :---- |
| KernelMaxPool3DWithIdx | 1 | 1103.2us | 966.28us | 14.16% |