@JamesLim-sy 2022-09-07
I think it is better to list the perf value with a table.